requirements.txt
numpy>=1.20.2
matplotlib>=3.3.3
rx>=3.2.0
scipy>=1.6.0
sounddevice>=0.4.2
einops>=0.3.0
tqdm>=4.64.0
pandas>=1.4.2
torchaudio>=0.10,<1.0
pyannote.core>=4.4
pyannote.database>=4.1.1
pyannote.metrics>=3.2
optuna>=2.10
websockets>=10.3
